requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain_openai
streamlit
pytest
